requirements.txt
absl_py==1.4.0
flax==0.6.4
jax==0.4.4
ml_collections==0.1.1
numpy==1.23.5
optax==0.1.4
scikit_learn==1.3.2
scipy==1.12.0
